Added support to set the fiddle name, and to run existing fiddles. Fixed the template coffeescript to actually work :)

+++ fiddle.sh
@@ -3,18 +3,32 @@
 if [ "$FIDDLEDIR" == "" ]; then
     FIDDLEDIR=/tmp
 fi
-FIDDLE=$FIDDLEDIR/`python -c "import random; print hex(int(str(random.random())[2:10]))[2:]"`
+
+if [ "$1" == "" ]; then
+    FIDDLE=$FIDDLEDIR/`python -c "import random; print hex(int(str(random.random())[2:10]))[2:]"`
+else
+    FIDDLE=$FIDDLEDIR/$1
+fi
+
 VIM=mvim
 
-echo Creating fiddle to $FIDDLE
+if [ -e $FIDDLE ]; then
+    echo "Fiddle already exists, running it..."
+    (set -e;
+      cd $FIDDLE;
+      $VIM -S fiddle.vim
+      python server.py)
+else
+    echo Creating fiddle to $FIDDLE
 
-(set -e;
-    mkdir $FIDDLE;
-    cp template/* $FIDDLE;
-    cd $FIDDLE;
-    haml index.haml index.html;
-    sass main.scss main.css;
+    (set -e;
+        mkdir $FIDDLE;
+        cp template/* $FIDDLE;
+        cd $FIDDLE;
+        haml index.haml index.html;
+        sass main.scss main.css;
 
-    $VIM -S fiddle.vim;
-    python server.py
-)
+        $VIM -S fiddle.vim;
+        python server.py
+    )
+fi
